mcp_server_postgres/requirements.txt
modelcontextprotocol
psycopg2-binary
fastapi
uvicorn[standard]
sqlalchemy
alembic
python-dotenv
